Refactoring Managers
# New
- Worlds now have an explicit disposal function.
- Entities now track whether they have started.
# Changed
- Renaming `Managers` to `Manager` in preparation for an engine design refactoring,
# Removed
- Removed test function from `MainApp`.
# Fix
- Fixed `three` being a part of ThreeWorld's dependency array.
- Fixed Worlds incorrectly disposing entities when
